split syscall
