+++ mkfs.c
@@ -267,6 +267,9 @@ iappend(uint inum, void *xp, int n)
   // printf("append inum %d at off %d sz %d\n", inum, off, n);
   while(n > 0){
     fbn = off / BSIZE;
+    if(fbn >= MAXFILE) {
+        printf("%d >= %lu\n", fbn, MAXFILE);
+    }
     assert(fbn < MAXFILE);
     if(fbn < NDIRECT){
       if(xint(din.addrs[fbn]) == 0){
