+++ usertests.c
@@ -1737,14 +1737,6 @@ void argptest()
   printf(1, "arg test passed\n");
 }
 
-unsigned long randstate = 1;
-unsigned int
-rand()
-{
-  randstate = randstate * 1664525 + 1013904223;
-  return randstate;
-}
-
 int
 main(int argc, char *argv[])
 {
